requirements.txt
ultralytics
flask
opencv-python
numpy
torch
torchvision
matplotlib
Pillow